sandbox-suite/apps/python-flask/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
ddtrace==2.6.0

# Intentionally vulnerable dependencies (SCA findings)
# Jinja2 3.1.2 still has CVE-2024-22195 (XSS), compatible with Flask 3.0.0
pyyaml==5.3.1
Jinja2==3.1.2
requests==2.25.0
Pillow==9.5.0
